[macros] flow macro based on nozzle size and line width
During a test with 0.6 nozzle, i saw in consistencies with line width vs actual print measurement.
So rather than potentially forgetting to set the proper amount in the slicer, i made a little macro to do it in via klipper.
So Slicer can stay at 100% flow and i wont have to make multiple profiles per line width if I choose to change it