[macros] enable arcs g0/g1 g2/g3
Was having too much zits/blobs on the start/stops, but also had on regular curves bt not on the straight walls of a temp tower. So it had to be curve related.

Even though klipper is supposed to be faster an not need "arcwelder", maybe with 2 cameras and 2 instances the comms are slower? Because it was still happening when the seam was aligned, and when lifting up the the nozzle to reduce pressure (even tho the first 2 prints were squished too much by .3 mm) it showed the same after correcting the z offset. And then once arc welder was enabled, then those zits went away.