DISABLE ArcWelder for klipper
Found all the tiny G1/2/3/4 conversions to linear lines it was doing was causing the MCU to crash.

Been having this issue where klipper shutdown due to MCU "Timer too close" error message.
 Get shutdown log here:

   523  cd ~/klipper/scripts/
   524  ./logextract.py ~/Ender3Pro_data/logs/klippy.log
   525  cd ~/Ender3Pro_data/logs/
   526  ls -l

Downloading the klipper.log and grpahing it here Graph Klipper Stats (sineos.github.io) dooesnt show MCU that busy, although bandwidth is high.

Would occur more often when Timelapse is enabled, though it still happens without it.
Removed all cameras and turned off obico and it still occurred.

Would happen quicker when running at a super fast speed (500%)

So it seemed to be related to movement.

Remembered that I recently enabled ArcWelder to deal with pimples on arcs.
Changed from 0.05 resolution to .1, still crashed.

Disabled arcwelder and running at 800% with timelapse and still no crashing.
So that seems to be the issue.

Will need to test pimples again but I think that occurs on "random seam", but for regular seam it should be okay. Otherwise might need to play with a safe arc value.
And I think it was also due to "alternate wall direction" not sure